Refactor onSelectExperiment handler
